+++ coroutines/src/commonMain/kotlin/dev/inmo/micro_utils/coroutines/DoInContext.kt
@@ -1,8 +1,6 @@
 package dev.inmo.micro_utils.coroutines
 
-import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.withContext
+import kotlinx.coroutines.*
 
 suspend inline fun <T> doInUI(noinline block: suspend CoroutineScope.() -> T) = withContext(
     Dispatchers.Main,
@@ -10,9 +8,5 @@ suspend inline fun <T> doInUI(noinline block: suspend CoroutineScope.() -> T) = 
 )
 suspend inline fun <T> doInDefault(noinline block: suspend CoroutineScope.() -> T) = withContext(
     Dispatchers.Default,
-    block
-)
-suspend inline fun <T> doInIO(noinline block: suspend CoroutineScope.() -> T) = withContext(
-    Dispatchers.IO,
     block
 )